refactor(bridge): Split environment-details component
