+++ bridge/client/app/_components/ktb-evaluation-details/ktb-evaluation-chart/ktb-evaluation-chart-legacy/ktb-evaluation-chart-legacy.component.html
@@ -1,0 +1,162 @@
+<div class="chart-container m-0" (click)="$event.stopPropagation()">
+  <button
+    dt-button
+    variant="secondary"
+    class="button-show-more-slo"
+    *ngIf="!isHeatmapExtended && isHeatmapExtendable && chartType === 'heatmap'"
+    (click)="toggleHeatmap()"
+    [style.top.px]="heatmapHeight + 20"
+    [style.width.px]="heatmapWidth - 8"
+  >
+    <dt-icon name="sorter2-down"></dt-icon>
+    Show all SLIs
+  </button>
+  <button
+    dt-button
+    variant="secondary"
+    class="button-show-more-slo"
+    *ngIf="isHeatmapExtended && isHeatmapExtendable && chartType === 'heatmap'"
+    [style.top.px]="heatmapHeight + 20"
+    [style.width.px]="heatmapWidth - 8"
+    (click)="toggleHeatmap()"
+  >
+    <dt-icon name="sorter2-up"></dt-icon>
+    Collapse to 10 SLIs
+  </button>
+  <dt-chart
+    class="mr-2"
+    #heatmapChart
+    *ngIf="chartType === 'heatmap'"
+    uitestid="keptn-evaluation-details-chartHeatmap"
+    [options]="_heatmapOptions"
+    [series]="heatmapSeries"
+    (seriesVisibilityChange)="seriesVisibilityChanged($event)"
+  >
+    <dt-chart-tooltip>
+      <ng-template let-tooltip>
+        <dt-key-value-list
+          style="min-width: 100px"
+          *ngIf="tooltip.point.point | toType: HeatmapPointClass as dataPoint"
+        >
+          <dt-key-value-list-item>
+            <dt-key-value-list-key> Value</dt-key-value-list-key>
+            <dt-key-value-list-value>
+              {{ dataPoint.value | truncateNumber: 2 }}
+            </dt-key-value-list-value>
+          </dt-key-value-list-item>
+
+          <ng-container *ngIf="dataPoint.sliInfo as info">
+            <dt-key-value-list-item>
+              <dt-key-value-list-key> Total passed SLIs</dt-key-value-list-key>
+              <dt-key-value-list-value>
+                {{ info.passCount }}
+              </dt-key-value-list-value>
+            </dt-key-value-list-item>
+
+            <dt-key-value-list-item>
+              <dt-key-value-list-key> Total warning SLIs</dt-key-value-list-key>
+              <dt-key-value-list-value>
+                {{ info.warningCount }}
+              </dt-key-value-list-value>
+            </dt-key-value-list-item>
+
+            <dt-key-value-list-item>
+              <dt-key-value-list-key> Total failed SLIs</dt-key-value-list-key>
+              <dt-key-value-list-value>
+                {{ info.failedCount }}
+              </dt-key-value-list-value>
+            </dt-key-value-list-item>
+
+            <span class="key-value-parent"> Threshold </span>
+            <div class="ml-2">
+              <dt-key-value-list-item>
+                <dt-key-value-list-key>pass</dt-key-value-list-key>
+                <dt-key-value-list-value>
+                  <span [class.error]="info.fail" [textContent]="info.thresholdPass"></span>
+                </dt-key-value-list-value>
+              </dt-key-value-list-item>
+
+              <dt-key-value-list-item>
+                <dt-key-value-list-key>warning</dt-key-value-list-key>
+                <dt-key-value-list-value>
+                  <span
+                    [class.warning]="info.warn && !info.fail"
+                    [class.error]="info.fail"
+                    [textContent]="info.thresholdWarn"
+                  ></span>
+                </dt-key-value-list-value>
+              </dt-key-value-list-item>
+            </div>
+          </ng-container>
+
+          <ng-container *ngIf="dataPoint.data">
+            <dt-key-value-list-item>
+              <dt-key-value-list-key> Score</dt-key-value-list-key>
+              <dt-key-value-list-value>
+                {{ dataPoint.data.score | truncateNumber: 2 }}
+              </dt-key-value-list-value>
+            </dt-key-value-list-item>
+
+            <dt-key-value-list-item>
+              <dt-key-value-list-key> Key SLI</dt-key-value-list-key>
+              <dt-key-value-list-value>
+                {{ dataPoint.data.keySli ? 'true' : 'false' }}
+              </dt-key-value-list-value>
+            </dt-key-value-list-item>
+
+            <span class="key-value-parent">Pass Criteria</span>
+            <div class="ml-2">
+              <dt-key-value-list-item *ngFor="let target of dataPoint.data.passTargets">
+                <dt-key-value-list-key> {{ target.criteria }} </dt-key-value-list-key>
+                <dt-key-value-list-value>
+                  <span [class.error]="target.violated"> {{ formatNumber(target.targetValue) }} </span>
+                </dt-key-value-list-value>
+              </dt-key-value-list-item>
+            </div>
+
+            <span class="key-value-parent"> Warning Criteria </span>
+            <div class="ml-2">
+              <dt-key-value-list-item *ngFor="let target of dataPoint.data.warningTargets">
+                <dt-key-value-list-key> {{ target.criteria }} </dt-key-value-list-key>
+                <dt-key-value-list-value>
+                  <span [class.error]="target.violated"> {{ formatNumber(target.targetValue) }} </span>
+                </dt-key-value-list-value>
+              </dt-key-value-list-item>
+            </div>
+          </ng-container>
+        </dt-key-value-list>
+      </ng-template>
+    </dt-chart-tooltip>
+  </dt-chart>
+
+  <dt-chart
+    *ngIf="chartType === 'chart'"
+    uitestid="keptn-evaluation-details-chartSeries"
+    [options]="_chartOptions"
+    [series]="_chartSeries"
+    (seriesVisibilityChange)="seriesVisibilityChanged($event)"
+  >
+    <dt-chart-tooltip>
+      <ng-template let-tooltip>
+        <p *ngIf="getEvaluationFromPoint(tooltip.points) as tooltipEvaluation">
+          SLO evaluation of <span [textContent]="tooltipEvaluation.data.teststrategy"></span> test from
+          <span
+            class="m-0 mt-1 mb-1"
+            *ngIf="tooltipEvaluation.time"
+            [textContent]="tooltipEvaluation.time | amDateFormat: dateUtil.getDateTimeFormat()"
+          ></span>
+        </p>
+        <dt-key-value-list style="min-width: 100px">
+          <dt-key-value-list-item *ngFor="let data of filterPoints(tooltip.points)">
+            <dt-key-value-list-key>
+              {{ data.series.name }}
+            </dt-key-value-list-key>
+            <dt-key-value-list-value>
+              {{ data.point.y | truncateNumber: 2 }}
+            </dt-key-value-list-value>
+          </dt-key-value-list-item>
+        </dt-key-value-list>
+      </ng-template>
+    </dt-chart-tooltip>
+  </dt-chart>
+</div>
