+++ bridge/client/app/_components/ktb-evaluation-details/ktb-evaluation-chart/ktb-evaluation-chart-legacy/ktb-evaluation-chart-legacy.component.scss
@@ -1,0 +1,31 @@
+:host ::ng-deep {
+  path.highlight-primary {
+    stroke-width: 2;
+    stroke: rgb(109, 109, 109);
+    fill: rgba(109, 109, 109, 0.4);
+  }
+
+  path.highlight-secondary {
+    stroke-width: 1;
+    stroke-dasharray: 5, 5;
+    stroke: rgb(137, 137, 137);
+    fill: rgba(137, 137, 137, 0.2);
+  }
+
+  .dt-chart-pane {
+    overflow: visible !important;
+  }
+}
+
+.chart-container {
+  position: relative;
+}
+
+.key-value-parent {
+  font-size: 12px;
+}
+
+.button-show-more-slo {
+  position: absolute;
+  z-index: 1;
+}
