+++ bridge/client/app/_components/ktb-evaluation-details/ktb-evaluation-details.component.scss
@@ -1,28 +1,6 @@
 @import '../../../variables';
 @import '~@dynatrace/barista-components/core/src/style/variables';
 
-.key-value-parent {
-  font-size: 12px;
-}
-
-::ng-deep {
-  path.highlight-primary {
-    stroke-width: 2;
-    stroke: rgb(109, 109, 109);
-    fill: rgba(109, 109, 109, 0.4);
-  }
-
-  path.highlight-secondary {
-    stroke-width: 1;
-    stroke-dasharray: 5, 5;
-    stroke: rgb(137, 137, 137);
-    fill: rgba(137, 137, 137, 0.2);
-  }
-
-  .dt-chart-pane {
-    overflow: visible !important;
-  }
-}
 .dt-button + .dt-button {
   fill: $cta-color-active;
   margin-left: $gap-normal;
@@ -30,13 +8,4 @@
 
 .wide-input {
   min-width: 440px;
-}
-
-.chart-container {
-  position: relative;
-}
-
-.button-show-more-slo {
-  position: absolute;
-  z-index: 1;
 }
