+++ bridge/client/app/_views/ktb-evaluation-view/ktb-evaluation-view.component.html
@@ -65,7 +65,7 @@
           <ktb-event-item-detail>
             <ktb-evaluation-details
               *ngIf="evaluation.data.evaluation"
-              [evaluationData]="evaluation"
+              [evaluation]="evaluation"
             ></ktb-evaluation-details>
           </ktb-event-item-detail>
         </ktb-event-item>
