+++ bridge/client/app/_views/ktb-service-view/ktb-service-details/ktb-service-details.component.html
@@ -69,7 +69,7 @@
       <ktb-event-item [showChartLink]="true" [event]="evaluationEvent" [showTime]="false" [showLabels]="false">
         <ktb-event-item-detail>
           <ktb-evaluation-details
-            [evaluationInfo]="{
+            [evaluationData]="{
               evaluation: evaluationEvent,
               shouldSelect: deploymentStage.hasEvaluation
             }"
