+++ bridge/shared/models/trace.ts
@@ -22,7 +22,6 @@ export interface IEvaluationData {
   score_state?: string;
   compare_with?: string;
   include_result_with_score?: string;
-  number_of_missing_comparison_results?: number;
   sloFileContentParsed?: true; // can be undefined (default) or true
   sloObjectives?: ISloObjectives[];
 }
